feat: add Final Boss death callback
Will need to trigger the change to the endgame scene once the boss is dead

+++ Game/Scripts/EnemyDeathScript.cpp
@@ -65,7 +65,7 @@ void EnemyDeathScript::ManageEnemyDeath()
 	}
 
 	DisableEnemyActions();
-	float enemiesLeft = App->GetModule<ModulePlayer>()->GetEnemiesToDefeat() - 1;
+	float enemiesLeft = static_cast<float>(App->GetModule<ModulePlayer>()->GetEnemiesToDefeat()) - 1.f;
 	App->GetModule<ModulePlayer>()->SetEnemiesToDefeat(enemiesLeft);
 }
